requirements.txt
aiohappyeyeballs==2.3.4
aiosignal==1.3.1
amqp==5.2.0
asgiref==3.8.1
attrs==24.1.0
beautifulsoup4==4.12.3
billiard==4.2.0
celery==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.6.1
Django==5.0.7
djangorestframework==3.15.2
drf-spectacular==0.27.2
frozenlist==1.4.1
gevent==24.2.1
greenlet==3.0.3
h11==0.14.0
idna==3.7
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kombu==5.3.7
multidict==6.0.5
outcome==1.3.0.post0
packaging==24.1
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
redis==5.0.8
referencing==0.35.1
requests==2.32.3
rpds-py==0.19.1
selenium==4.23.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.5.1
trio==0.26.1
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.9.4
zope.event==5.0
zope.interface==7.0
